[Athlete Chart] Scout Details - Improve load time for graphs.  (#435)
* fix: improved loading time for graphs

* refactor: move athlete page tooltip to separate file

* refactor: changed how the set was initialized
