+++ lib/pages/athlete/AthletePage.dart
@@ -1,4 +1,5 @@
 import 'package:ax_dapp/pages/athlete/bloc/AthletePageBloc.dart';
+import 'package:ax_dapp/pages/athlete/components/AthletePageTooltip.dart';
 import 'package:ax_dapp/pages/athlete/components/BuildLongChart.dart';
 import 'package:ax_dapp/pages/athlete/components/BuildShortChart.dart';
 import 'package:ax_dapp/pages/athlete/components/Buttons.dart';
@@ -1031,38 +1032,13 @@ class _AthletePageState extends State<AthletePage> {
                                 Row(
                                   mainAxisAlignment: MainAxisAlignment.end,
                                   children: <Widget>[
-                                    scoutPageToolTip(),
+                                    athletePageToolTip(),
                                   ],
                                 ),                                 
                               ]))
                     ])),
           ],
         ));
-  }
-
-  Widget scoutPageToolTip() {
-    return Tooltip(
-      height: 20,
-      padding: EdgeInsets.all(20),
-      preferBelow: true,
-      decoration: BoxDecoration(
-          color: Colors.grey[800],
-          borderRadius: BorderRadius.circular(50)),
-      richMessage: TextSpan(
-        children: <TextSpan>[
-          TextSpan(
-            text: "Buy / Sell individual APT's at their Market Price\n",
-            style: TextStyle(color: Colors.grey[400], fontSize: 18)
-          ),
-          TextSpan(
-            text: "Mint / Redeem APT Pairs for their Book Value",
-            style: TextStyle(color: Colors.grey[400], fontSize: 18)
-          ),
-        ],
-      ),
-      child:
-          Icon(Icons.info_outline_rounded, color: Colors.grey, size: 25),
-    );
   }
 
   Widget statsSide(
