+++ lib/pages/athlete/components/AthletePageTooltip.dart
@@ -1,0 +1,25 @@
+  import 'package:flutter/material.dart';
+  Widget athletePageToolTip() {
+    return Tooltip(
+      height: 20,
+      padding: EdgeInsets.all(20),
+      preferBelow: true,
+      decoration: BoxDecoration(
+          color: Colors.grey[800],
+          borderRadius: BorderRadius.circular(50)),
+      richMessage: TextSpan(
+        children: <TextSpan>[
+          TextSpan(
+            text: "Buy / Sell individual APT's at their Market Price\n",
+            style: TextStyle(color: Colors.grey[400], fontSize: 18)
+          ),
+          TextSpan(
+            text: "Mint / Redeem APT Pairs for their Book Value",
+            style: TextStyle(color: Colors.grey[400], fontSize: 18)
+          ),
+        ],
+      ),
+      child:
+          Icon(Icons.info_outline_rounded, color: Colors.grey, size: 25),
+    );
+  }
